tracing: Handle resuming data source with async stop after reconnection
Scenario:
1. The tracing service disconnects while a data source is active.
2. TracingMuxerImpl asks the data source to stop.
3. The data source postpones the stop.
4. The tracing service reconnects.
5. It tries to reactivate the data source.

TracingMuxerImpl now refuses to reactivate the data source, because it
says that the same data source instance with the same config is present.

Actually the data source instance that's already present is the one that
is being stopped and actually references the old producer (therefore it
cannot write any data).

This commit fixes the problem by ignoring the old being-stopped data
source instance.

Found when investigating a similar problem for crbug.com/1495261

Change-Id: I68856486a0397d05d00dc1459bf6af0de1e99fa5