Fix rendering of flows to descendants
If a flow connects a slice to its descendant, currently the rendering
looks unpleasant, especially if it repeats multiple times.

Explicitly detect this case and start the rendering the flow from the
start of the ancestor (instead from the end) in this case.

This patch also adds auto-included stdlib/prelude module and drops the requirement for the stdlib functions to be prefixed with the module name.

Change-Id: I35af81ebbc51de572f0e043eac0abc905c68bca6