tracing: Call FinishTracePacket in the TrackEvent data source
Before this change, the last event emitted for each thread was not
visible until the thread issued a Flush. Since TrackEvent can be used
from many threads and not all of them can emit a Flush, this meant that
the last event was always lost.

This fixes the problem by explicitly calling FinishTracePacket.

Bug: 162206162
Change-Id: Ifc0d1c662679708a2f802fb5074e046e48f35128