trace_buffer: Remove wrong check in WriteChunkRecord
The CHECK is correct most of the times, but, in case of writing a chunk
that's already been scraped it might not be.

The problem is that `size_to_end()` is `end() - wptr_` (with an
underscore). In case of copying an already scraped chunk, `wptr`
(without an underscore) is different than `wptr_` (with an underscore).
The CHECK thinks that the new record would not fit into the buffer,
because `wptr_` is almost at the end, but really we're writing into
`wptr`.

I think the right thing to do is to remove the CHECK. The callers of
WriteChunkRecord fill `record.size` in a trivially correct way, I think.

Bug: 311619372
Change-Id: Ie517d2bda8a335ce6b7fb2af10532efc55c56103