UI: clean up Modal dialog
The way the modal dialog was implemented was incredibly
dirty and overengineered. The core problem with modal dialog
is that it's "logically" a children of some mithril component
but "physically" needs to be rendered in the root of the (v)DOM
in order to be able to overlay on top of everything (although in
some rare cases it can be rendered in a details panel, like the
flamegraph does).
I don't know why I ended up introducing all those indirection
layers in the past. Clearly the right thing to do was just
taking as argument a factory function that returns
mithril components (thanks Steve!).
On the nice side, this refactoring keeps the API of showModal
almost untouched, at the cost of supporting both
(1) static content, where the input is an m.Children
(2) dynamic content via a factory.

Test: manual: flamegraph with incomplete graph
Test: manual: adding target to recording v2
Test: manual: adding a column to pivot table v2
Test: manual: '?' help menu
Test: manual: postmessage from an untrusted origin
Test: manual: inducing a crash

Change-Id: Ibff8309bc7e28d161ecf21c7e55e11d4e000c592