traced: report buffer id in trace writer stats
I recently introduced trace writer stats. In the previous
CL I was merging writer stat histograms across all buffers.
That makes little sense because in reality, it's extremely hard
for a writer to write into two different buffers. That merging
was NOT technically wrong, just useless.
On the other hand, is very beneficial to report the buffer index
for each writer, so we can better identify data sources and who
wrote into what.

Test: perfetto_unittests --gtest_filter=TracingServiceImplTest.TraceWriterStats
Bug: 313851266
Change-Id: If09e14b732404823a6e3d3dc61b627697f1a91da