[ui] Add "@typescript-eslint/strict-boolean-expressions" eslint rule
This rule:
 - Prohibits using numbers in boolean expressions - fixes the classic
   bug where 0 is considered falsy.
 - Prohibits use of always true/false values in boolean expressions.

Note: This rule can be configured to catch even more cases, but
it's initially reined in to avoid too much disruption.

Aside from adding the rule, this CL also adds some tweaks to get rules
working that require type information to run, and also adds a new npm
script `npm run lint` which runs eslint over the entire codebase.

It's expected that a lot of CLs will face presubmit issues as a result
of this change - these are addressed in a separate CL.

Bug: https://github.com/google/perfetto/issues/670
Change-Id: I5634765c3e9516267be4b8c59fc6f9f6db1800b8