Don't call OnStop twice on service disconnect
Calling OnStop twice in a row can lead to inconsistencies in the
datasource state (see crbug.com/1434211 for the example). One situation
when this can happen is when the producer is disconnected from the
service while async stop is in progress.

This CL makes sure we don't call OnStop twice in this case.

Bug: chromium:1434211
Change-Id: Ib38acaff1e01c2fc17c341f3195b5f814024d832