ftrace_stats: change "bytes_read" to the more accurate "bytes" + assorted fixups
I noticed that ftrace_stats.proto and derived trace processor entry
have a misleading entry. The "bytes_read" field represents the size of
valid tracing data in the buffer, not the data that's been read by
userspace. Hence a small renaming pass. I did keep the proto field name
the same to limit scope, and because the accurate name ("bytes") would
be reserved (so the new name wouldn't be exact either).

Also small simplification to ftrace_stats lifecycle in traced_probes.
It's still a bit convoluted, but less so.

Change-Id: I383021a9413db193dcbfc8d6e8e2ec287f27148b