trace_processor: refactor and simplify metrics function code
This CL works on simplifying some of the code for the metrics SQL
functions. This prepares for changing the implementation of the
RepeatedField function in a followup CL.

While doing this, since it's very helpful, also add some gtest matchers
for status OK/error checking which makes the tests nicer.

Change-Id: Icf1ad1417eab5e3d309f00fd340a22ba874ca7d2