tp: simplify StaticTableFunction API significantly
This CL removes the ValidateConstraints API method and significantly
simplifies the ComputeTable method on StaticTableFunction.

Back when StaticTableFunctions were actually DynamicTables, this more
complex API made more sense as we wanted to expose the full power of
SQLite's virtual table API. However, in practice, with one exception,
we never actually used this power. Most of the changes, with the
exception, are fully mechanical and should have no impact on anything.

The one exception is the flamegraph function: in this case, we were
actually making use of some of the weirder features (partial constraint
setting, non-equality constraints) which require changing the API of
flamegraph to deal with this.

As flamegraph is an experimental API, we can make this change without
any sort of deprecation happening.

Change-Id: I553b10191f4e6b99707f482f125374d3c91aaa22