AndroidNetwork plugin: ensure battery_stats metric is run before referencing its output
For users that disable the "run metric on load" UI flags, the network
plugin results in an error. The UI is fine after dismissing the error
dialog, but let's get rid of the dialog at the cost of running the
metric anyway.

Imo this is a stopgap, we shouldn't even run plugins on traces that
don't contain the relevant events (such as running android plugins on
desktop linux traces).

Bug: 322498817
Change-Id: Ic91750f92bfc0d4623c630c6f16a9f5a8e49abd1