tp: fix naming of debugging function
Make it clearer that the column::name is for debugging only by using
the more standardised "DebugString" naming (see e.g. protobuf library
which does this for all autogenerated proto classes).

Also switching to std::string allows us to adding more debugging info
in the future which is not static (e.g. the size of vectors etc).

Change-Id: I08200642af1fb3c1b1417ddad36eb8c30f14e2bc