Refactor StackProfile related classes.
At a high level this patch splits SequenceStackProfileTracker into a
class that deals with interned data (StackProfileSequenceState) and
another one that deals with stacks stored in the db
(StackProfileTracker).

Also HeapProfileTracker currently tracks per sequence data for Profile
packets so it has been renamed to ProfilePacketSequenceState and
rewritten so that there is one instance per sequence.

In the future there will be other ways of adding mappings, frames and
callstacks (e.g. perf trace files, or from the v8 data source). Having a
dedicated API to just intern such data into the DB will make that easier
and cleaner.

This cl also ties the lifetime of these state instances to the
sequence state generation. Since https://r.android.com/2897485 there is
no guarantee that a iid will refer to the same object across incremental
state resets. Thus it makes sense to keep these state objects attached
to the generation instead of the PacketSequenceState.

Bug: b/283794416
Change-Id: I97b1bfe27ea36d6e67ea80a1f0267457cbda2d35