service: rename --clone-skip-filter to --clone-for-bugreport
The previous CL introduced a --clone-skip-filter option to skip
the filter when cloning traces for bugreport.
We need a slighly different semantic here that is: "skip the filter
AND mark this trace as a bugreport trace (for flushes)".
Renaming the cmdline argument and adding another protocol boolean
to reflect that.

Bug: 321196572
Change-Id: I00d5ed4c19546992e9581c74f658dd8ce5b03118